RPGgame (THE JOURNEY)
A simple game written in C language as a part of an assignment. The
purpose of the game is to encourage players to using programming
languages. The game is played by writing codes to interact with the
surroundings. This is game is not commercial, it is only for educational
purposes. It was designed with a group team (AKRAM, TONI, RACHEAL,
MINETA, OMAR) as an assignment for university of Bristol The code is in
C, SDL is used for graphics

The concept is to find your way from your house to the castle
Puzzles will face you during playing
all actions should be done by writing commands like (Action.Open(BOX),
Action.Take(ITEM)....etc)
Further Explanation is in the gameplay.